Optimized BP code for regenerating and draining resilience. Instead of calling unneeded variables and timers and destroying them every time block is called, it has a continuous timer that checks every second if the draining/regenerating conditions are met (regen if not full resilience or blocking, etc). As a result, several booleans and timer variables are removed.
Moved BlockHandler to subclass instead of Base Character BP (and as a result, will not be included in C++ class) due to possibly necessity to modify block for child classes.

Modified CooldownDecrement to follow new naming scheme (VariableName instead of camelcase variableName) along with changing it to match the original function instead of having a different name. This is because ultimately we will be removing the functions in the BP counterpart, so changing it now would just be a waste of effort.